refactor(article): Implement feedback of Quinn, move fn
